requirements.txt
Pillow
matplotlib
numpy
pytorch>=0.4.0
seaborn
torchvision
